+++ app/assets/javascripts/folio/console/base.js
@@ -34,6 +34,7 @@
 //= require folio/console/_modal-lazyload
 //= require folio/console/_modal-html-scroll
 //= require folio/console/_cocoon-set-position
+//= require folio/console/_cocoon-prompt-file-input
 //= require folio/console/_selectize
 //= require folio/console/_simple-form-error-scrollto
 //= require folio/console/_tabs
@@ -41,12 +42,12 @@
 //= require folio/console/tagsinput/tagsinput
 //= require folio/console/nested_model_controls/nested_model_controls
 //= require folio/console/boolean_toggle/boolean_toggle
-//= require folio/console/index_position_buttons/index_position_buttons
 //= require folio/console/single_file_select/single_file_select
 //= require folio/console/react_picker/react_picker
 //= require folio/console/menu_tree/menu_tree
 //= require folio/console/file_list/file_list
 //= require folio/console/index/checkboxes/checkboxes
+//= require folio/console/index/position_buttons/position_buttons
 //= require folio/console/modules/_layout
 //= require folio/console/modules/_dirty-forms