+++ app/controllers/folio/console/accounts_controller.rb
@@ -9,17 +9,17 @@ class Folio::Console::AccountsController < Folio::Console::BaseController
 
   def create
     @account = Folio::Account.create(account_params)
-    respond_with @account, location: { action: :index }
+    respond_with @account
   end
 
   def update
     @account.update(account_params)
-    respond_with @account, location: { action: :index }
+    respond_with @account
   end
 
   def destroy
     @account.destroy
-    respond_with @account, location: { action: :index }
+    respond_with @account
   end
 
   private