+++ config/initializers/show_for.rb
@@ -68,6 +68,13 @@ class ShowFor::Builder
     toggle(:published)
   end
 
+  def position_controls
+    attribute(:position) do
+      template.cell('folio/console/index/position_buttons', object).show
+                                                                   .html_safe
+    end
+  end
+
   def type
     attribute(:type) do
       object.class.model_name.human
@@ -101,7 +108,7 @@ class ShowFor::Builder
       if object.persisted?
         if object.visit.present?
           template.link_to(object.visit.to_label,
-                           template.url_for([:console, object.visit]))
+                           template.controller.url_for([:console, object.visit]))
         end
       end
     end
@@ -130,7 +137,7 @@ class ShowFor::Builder
             content = object.public_send(attr)
           end
 
-          url = template.url_for(url_for_args)
+          url = template.controller.url_for(url_for_args)
           template.link_to(content, url)
         end
       end