+++ lib/folio/engine.rb
@@ -40,6 +40,9 @@ module Folio
     config.folio_pages_ancestry = false
     config.folio_console_locale = :cs
     config.folio_console_dashboard_redirect = :console_pages_path
+    config.folio_console_sidebar_prepended_link_class_names = []
+    config.folio_console_sidebar_appended_link_class_names = []
+    config.folio_console_sidebar_runner_up_link_class_names = []
 
     initializer :append_migrations do |app|
       unless app.root.to_s.match? root.to_s